Rename `multimodalmaestro` to `maestro`.
